Fix pre-select corner case for parent class reference fields
Fixed a bug that occurred when using *pre-select* to load a Record containing a reference field whose **declared** type is a parent class of a descendant with additionally defined fields and the stored value for that field is an instance of that child class. In this case, the pre-select logic did not load data for the additionally defined fields in the child class, which resulted in unexpected `NullPointerException` regressions or an overall inability to load those Records if the additional fields were annotated as `Required`.